fix(tests): modify test behavior in feed screen to implement new notifications and remove duplicate code
