--- app/src/androidTest/java/com/android/ootd/ui/feed/FeedScreenExtraCoverageTest.kt
@@ -1,72 +1,0 @@
-package com.android.ootd.ui.feed
-
-import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.test.onNodeWithContentDescription
-import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onRoot
-import androidx.compose.ui.test.performClick
-import com.android.ootd.model.posts.OutfitPost
-import org.junit.Assert.assertTrue
-import org.junit.Rule
-import org.junit.Test
-
-/**
- * Extra coverage tests for FeedScreen. These don't test business logic — they just trigger
- * currently unimplemented UI callbacks so Jacoco reports those lines as covered.
- */
-class FeedScreenExtraCoverageTest {
-
-  @get:Rule val composeTestRule = createComposeRule()
-
-  @Test
-  fun feedScreen_triggersTopBarButtons() {
-    var searchClicked = false
-    var notificationsClicked = false
-
-    composeTestRule.setContent {
-      FeedScreen(
-          onAddPostClick = {},
-          onSearchClick = { searchClicked = true },
-          onNotificationIconClick = { notificationsClicked = true })
-    }
-
-    // Wait for the UI to fully compose, especially for AccountIcon which uses a ViewModel
-    // The AccountIcon initializes a ViewModel that launches coroutines, so we need to give it time
-    composeTestRule.waitForIdle()
-
-    // Additional wait to ensure AccountIcon's ViewModel has initialized
-    Thread.sleep(100)
-    composeTestRule.waitForIdle()
-
-    // Simulate clicks on both icons
-    composeTestRule.onNodeWithContentDescription("Search").performClick()
-
-    // Use onNodeWithTag with assertion to ensure it exists before clicking
-    composeTestRule
-        .onNodeWithTag(FeedScreenTestTags.NAVIGATE_TO_NOTIFICATIONS_SCREEN)
-        .assertExists()
-        .performClick()
-
-    assertTrue(searchClicked)
-    assertTrue(notificationsClicked)
-  }
-
-  @Test
-  fun feedScreen_rendersOutfitPostCard_withFakeClick() {
-    // This test only exists to exercise the onSeeFitClick lambda inside OutfitPostCard
-    val fakePost =
-        OutfitPost(
-            postUID = "fake-id", ownerId = "user1", outfitURL = "https://example.com/fake.jpg")
-
-    composeTestRule.setContent {
-      // Directly render a single card to “touch” the lambda line
-      OutfitPostCard(
-          post = fakePost,
-          isBlurred = false,
-          onSeeFitClick = { /* fake click to mark as covered */})
-    }
-
-    // Just ensure the composable renders
-    composeTestRule.onRoot().assertExists()
-  }
-}
